fix: moved episode section to routes dir
