+++ src/components/index.ts
@@ -1,6 +1,7 @@
 export * from "./button";
 export * from "./card";
 export * from "./download-item";
+export * from "./episode-card";
 export * from "./poster-card";
 export * from "./problem-set-card";
 export * from "./team-member";
